newChicPrreprocessor/LocalFunctionProj/requirements.txt
azure-functions
tensorflow
Pillow
requests
